+++ src/main/java/com/streamr/client/protocol/message_layer/StreamMessageV30.java
@@ -59,7 +59,7 @@ public class StreamMessageV30 extends StreamMessage {
     }
 
     @Override
-    public int getSequenceNumber() {
+    public long getSequenceNumber() {
         return messageID.getSequenceNumber();
     }
